1 - changes from Aziz - 23.12.2019
1 - changes from Aziz - 23.12.2019

+++ UI/shaders/vshader.glsl
@@ -5,7 +5,7 @@ uniform mat4 uMVMatrix;
 
 in vec4 vV;
 in float vValue;
-in int vBlockIndex;
+in float vBlockIndex;
 
 out vec4 gVt;
 out vec4 gVp;
